made the algo more efficient, hell yeah!

+++ c3p_jax/utils.py
@@ -7,8 +7,14 @@ def place(M, block, r, c):
     M = M.at[r:r+rr, c:c+cc].set(block)
     return M
 
-def create_rho_schedule(n_iter: int, var_weights: jnp.ndarray, mult: float = 8, offset: int = 5, start_with_0_rho: bool = True) -> jnp.ndarray:
-    rho = jnp.outer(mult ** (jnp.arange(n_iter, dtype=jnp.float32) - offset), var_weights)
+def create_rho_schedule(
+    n_iter: int, 
+    var_weights: jnp.ndarray, 
+    mult: float = 8, 
+    offset: int = 5, 
+    start_with_0_rho: bool = True,
+) -> jnp.ndarray:
+    rho = jnp.outer(mult ** (jnp.arange(n_iter, dtype=var_weights.dtype) - offset), var_weights)
     if start_with_0_rho:
         rho = rho.at[0].set(1e-7)
     return rho
